backend/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
httpx
pydantic
huggingface_hub
openai
scikit-learn
pandas
numpy
python-multipart
google-genai
